backend/requirements.txt
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
# djangorestframework_simplejwt==5.5.1
pillow==12.0.0
psycopg==3.2.12
psycopg-binary==3.2.12
# pycparser==2.23
django-environ==0.12.0
gunicorn==23.0.0
whitenoise==6.11.0
yfinance==0.2.66
torch==2.8.0
torchvision==0.23.0
numpy==2.3.4
pandas==2.3.3
scikit-learn==1.7.2